requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.4.2
frozenlist==1.4.0
idna==3.6
magic-filter==1.0.12
motor==3.3.1
multidict==6.0.4
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
typing_extensions==4.8.0
yarl==1.9.3
